requirements.txt
openai>=1.0.0
google-cloud-aiplatform>=1.36.0
python-dotenv>=1.0.0
pillow>=10.0.0
requests>=2.31.0 